requirements.txt
PyMuPDF>=1.24.0
Pillow>=10.0.0
pytesseract>=0.3.10
openpyxl>=3.1.0
PyYAML>=6.0
